Both: Rework firmware/filesystem sizes.
The W build needs additional space for the wireless firmware.

This is painful since it makes these builds non-interchangable
as the filesystem will move/be overwritten.

As such, change *only* the W build and avoid changing the existing
build so updates don't delete user files.

Additionally rework the sizes so that it's clearer we're
reserving space for the firmware and not just making up numbers.